requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
